+++ src/app/App.tsx
@@ -1,7 +1,7 @@
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import { defaultArticleState, OptionType } from './constants/articleProps';
-import styles from './styles/index.module.scss';
+import { Article } from '../components/article/Article';
+import { ArticleParamsForm } from '../components/article-params-form/ArticleParamsForm';
+import { defaultArticleState, OptionType } from '../constants/articleProps';
+import styles from './App.module.scss';
 import { CSSProperties, useState } from 'react';
 
 export const App = () => {
